fix yaml: remove old command for icon (dangerous bug), typo in hyphen
I had a command of a task overwritten in the yaml file  with a wrong
path by accident, then the path for the executable did not exist, but
I did not get an error referring to this. Instead the error was that
the outputs were not found (because the executable that produces these
outputs was never executed^^). That is quite a dangerous bug as it does
produce a partially misleading error message. Need to check if this is
aiida-shell or workgraph related.