add(parameters): part 1: when kw and generic ParamSeries
the `when` keyword was added to address #40. As a result, keeping
track of dates in the TimeSeries class is not necessary anymore. Still a
more generic ParamSeries container needs to be there for future
parametrized tasks, so it was implemented instead of just ditching
TimeSeries and temporarily use regular dictionaries.