Fixes publish plubish-docs build (pygraphviz build failure)
One needs to install graphviz over the system package manager before
installing pygraphviz to prevent an automatic build that fails. This can
is achieved with `sudo apt-get install graphviz graphviz-dev`.

The `--force` was in `hatch run docs:deploy --force` was ignored. hatch
does not support arguments so we directly use it in the hatch-script
definition. Replacing `--force` to ``--no-history`.`
See https://github.com/pypa/hatch/discussions/988

We also add the `workflow_dispatch` event so we can run the workflow
manually.