Adds parsing of ShellTask and IconTask
The parameters specified in the yaml corresponding to specific task
plugins file are now passed to their corresponding classes in core.
ShellTask and IconTask. For that pydantic discrimantors are used.
See for more information
https://docs.pydantic.dev/latest/concepts/unions/#discriminated-unions-with-callable-discriminator

To test the new changes formatting of represented member variables has
been added to the formatting of `core.Task` objects.

Documentation to the `core.Plugin` has been added and the class has been
renamed to `core.TaskPlugin`.